Move search components into search folder (#65)
* Moves search components into search folder
* Rename components extensions to jsx
