+++ plugin-hrm-form/src/components/search/SearchForm/index.jsx
@@ -3,12 +3,12 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import SearchIcon from '@material-ui/icons/Search';
 
-import FieldText from './FieldText';
-import FieldSelect from './FieldSelect';
-import FieldDate from './FieldDate';
-import { Container, SearchFields, StyledSearchButton } from '../Styles/HrmStyles';
-import { withConfiguration } from '../ConfigurationContext';
-import { contextObject, searchFormType } from '../types';
+import FieldText from '../../FieldText';
+import FieldSelect from '../../FieldSelect';
+import FieldDate from '../../FieldDate';
+import { Container, SearchFields, StyledSearchButton } from '../../../Styles/HrmStyles';
+import { withConfiguration } from '../../../ConfigurationContext';
+import { contextObject, searchFormType } from '../../../types';
 
 const getField = value => ({
   value,
